refactor: 북마크 버튼 통합 (#302)
