+++ src/components/features/card/place-card.tsx
@@ -5,7 +5,7 @@ import { memo } from 'react';
 
 import PlaceImage from '@/components/commons/place-image';
 import { PlaceCardProps } from '@/types/card.type';
-import BookmarkButton from '../like/bookmark-button';
+import BookmarkButton from '@/components/commons/bookmark-button';
 
 /**
  * 장소 카드 컴포넌트
@@ -37,7 +37,7 @@ const PlaceCard = ({
         />
         {/* 북마크 버튼 */}
         <div className="absolute right-2 top-2">
-          <BookmarkButton isBookmarked={isBookmarked} placeId={placeId} />
+          <BookmarkButton placeId={placeId} />
         </div>
       </div>
